added validation for exemption type
