+++ src/components/ExemptionSelector.jsx
@@ -3,12 +3,14 @@ import { Field } from "formik";
 import { RadioButton } from "../common/RadioButton";
 import { ExemptionTypes } from "../common/Constants";
 import DateRangeSelector from "./DateRangeSelector";
+import { GetExemptionFormErrors } from "../common/ExemptionUtilities";
 
 const ExemptionSelector = props => {
   const {
     exemption: exemptionFromProps = {},
     onExemptionSave = () => {}
   } = props;
+  const [isFormDirty, setIsFormDirty] = useState(false);
   const [formErrors, setFormErrors] = useState([]);
   const [exemption, setExemption] = useState(exemptionFromProps);
 
@@ -16,23 +18,14 @@ const ExemptionSelector = props => {
     setExemption(exemptionFromProps);
   }, [exemptionFromProps]);
 
-  const getFormErrors = () => {
-    const activeFormErrors = [];
-    const { fromDate, toDate } = exemption;
-
-    if (!fromDate) {
-      activeFormErrors.push({ key: "fromDate", error: "From Date Required" });
+  useEffect(() => {
+    if (isFormDirty) {
+      setFormErrors(GetExemptionFormErrors(exemption));
     }
-
-    if (!toDate) {
-      activeFormErrors.push({ key: "toDate", error: "To Date Required" });
-    }
-
-    return activeFormErrors;
-  };
+  }, [isFormDirty, exemption]);
 
   const saveExemption = () => {
-    const errors = getFormErrors();
+    const errors = GetExemptionFormErrors(exemption);
 
     if (errors.length === 0) {
       onExemptionSave(exemption);
@@ -43,6 +36,7 @@ const ExemptionSelector = props => {
   };
 
   const handleExemptionTypeChange = ({ type, label }) => {
+    setIsFormDirty(true);
     setExemption({
       ...exemption,
       ...{ type, label }
@@ -50,6 +44,7 @@ const ExemptionSelector = props => {
   };
 
   const handleExemptionDateChange = ({ fromDate, toDate }) => {
+    setIsFormDirty(true);
     setExemption({
       ...exemption,
       ...{ fromDate, toDate }
@@ -57,6 +52,7 @@ const ExemptionSelector = props => {
   };
 
   const resetSelector = () => {
+    setIsFormDirty(false);
     setExemption({});
   };
